# artifacts/openapi/wecms-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeCMS API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/system/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "GET /api/v1/system/files",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PagedFileSummary"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "file-center",
        "x-wecms-permission": "sys:file:list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mimeType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "POST /api/v1/system/files",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FileMutationResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "file-center",
        "x-wecms-permission": "sys:file:upload",
        "x-wecms-audit": {
          "module": "file-center",
          "resource": "files",
          "action": "create"
        },
        "x-wecms-rate-limit": "file_upload_policy"
      }
    },
    "/api/v1/system/files/{id:long}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "DELETE /api/v1/system/files/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "file-center",
        "x-wecms-permission": "sys:file:delete",
        "x-wecms-audit": {
          "module": "file-center",
          "resource": "files",
          "action": "delete"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      },
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "GET /api/v1/system/files/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FileDetailDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "file-center",
        "x-wecms-permission": "sys:file:detail"
      }
    },
    "/api/v1/system/files/{id:long}/download": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "GET /api/v1/system/files/{id:long}/download",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "file-center",
        "x-wecms-permission": "sys:file:download"
      }
    },
    "/api/v1/system/files/{id:long}/preview": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "GET /api/v1/system/files/{id:long}/preview",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "file-center",
        "x-wecms-permission": "sys:file:download"
      }
    },
...
    "/api/v1/system/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "GET /api/v1/system/settings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PagedSettingSummary"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "configuration",
        "x-wecms-permission": "sys:setting:list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/system/settings/reload-cache": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "POST /api/v1/system/settings/reload-cache",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "configuration",
        "x-wecms-permission": "sys:setting:reload-cache",
        "x-wecms-audit": {
          "module": "configuration",
          "resource": "settings",
          "action": "reload-cache"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/settings/validate-ip-rules": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "POST /api/v1/system/settings/validate-ip-rules",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValidateIpRulesResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateIpRulesRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "configuration",
        "x-wecms-permission": "sys:setting:validate-ip-rules",
        "x-wecms-audit": {
          "module": "configuration",
          "resource": "settings",
          "action": "validate-ip-rules"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/settings/{key}": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "GET /api/v1/system/settings/{key}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SettingDetailDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "configuration",
        "x-wecms-permission": "sys:setting:detail"
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "PUT /api/v1/system/settings/{key}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SettingMutationResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "configuration",
        "x-wecms-permission": "sys:setting:update",
        "x-wecms-audit": {
          "module": "configuration",
          "resource": "settings",
          "action": "update"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResult": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "nullable": true
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "fieldErrors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "Object": {
        "type": "object",
        "additionalProperties": true
      },
...
      "PagedSettingSummary": {
        "type": "object",
        "required": [
          "records",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettingSummaryDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SettingSummaryDto": {
        "type": "object",
        "required": [
          "key",
          "valueType",
          "groupCode",
          "name",
          "isSensitive",
          "isSystem",
          "updatedAt"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": [
              "string",
              "null"
            ]
          },
          "valueType": {
            "type": "string"
          },
          "groupCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "isSensitive": {
            "type": "boolean"
          },
          "isSystem": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "SettingDetailDto": {
        "type": "object",
        "required": [
          "key",
          "valueType",
          "groupCode",
          "name",
          "isSensitive",
          "isSystem",
          "updatedAt"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": [
              "string",
              "null"
            ]
          },
          "valueType": {
            "type": "string"
          },
          "groupCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "isSensitive": {
            "type": "boolean"
          },
          "isSystem": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "UpdateSettingRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ValidateIpRulesRequest": {
        "type": "object",
        "required": [
          "rules"
        ],
        "properties": {
          "rules": {
            "type": "string"
          }
        }
      },
      "ValidateIpRulesResponse": {
        "type": "object",
        "required": [
          "valid"
        ],
        "properties": {
          "valid": {
            "type": "boolean"
          }
        }
      },
      "SettingMutationResponse": {
        "type": "object",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string"
          }
        }
      },
...
      "PagedFileSummary": {
        "type": "object",
        "required": [
          "records",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileSummaryDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FileSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "originalName",
          "fileExt",
          "mimeType",
          "sizeBytes",
          "sha256",
          "status",
          "createdBy",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "originalName": {
            "type": "string"
          },
          "fileExt": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "sha256": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FileDetailDto": {
        "type": "object",
        "required": [
          "id",
          "originalName",
          "fileExt",
          "mimeType",
          "sizeBytes",
          "sha256",
          "status",
          "createdBy",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "originalName": {
            "type": "string"
          },
          "fileExt": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "sha256": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateFileRequest": {
        "type": "object",
        "required": [
          "originalName",
          "mimeType",
          "sizeBytes",
          "sha256",
          "file"
        ],
        "properties": {
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "sha256": {
            "type": "string"
          },
          "policy": {
            "type": "string"
          },
          "file": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "FileMutationResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      }
    }
  }
}